Simple controls
Create a ControlWindow component that emulates a window within dwm. The
component has a dark and light theme with the former being the default.
Light theme can be enabled using the (boolean) `light` prop.

The ControlWindow component goes along with the ControlElementContainer
component. This component is suposed to contain elements that control
aspects of the screen (inputs, checkboxes, etc), it also comes with a
dark and light style much like ControlWindow.